+++ src/position.cpp
@@ -128,7 +128,7 @@ size_t xg_cached_node_length(id_t id, xg::XG* xgidx, LRUCache<id_t, Node>& node_
 }
 
 int64_t xg_cached_node_start(id_t id, xg::XG* xgidx, LRUCache<id_t, int64_t>& node_start_cache) {
-    pair<size_t, bool> cached = node_start_cache.retrieve(id);
+    pair<int64_t, bool> cached = node_start_cache.retrieve(id);
     if(!cached.second) {
         cached.first = xgidx->node_start(id);
         node_start_cache.put(id, cached.first);
